+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - support for NVML tunable parameters
 - option to simulate auto-tuning searches from existing cache files
 - Cupy backend to support C++ templated CUDA kernels
+- support for templated CUDA kernels using PyCUDA backend
+- documentation on tunable parameter vocabulary
 
 ## [0.3.2] - 2020-11-04
 ### Added